+++ src/components/RowOptions.js
@@ -1,0 +1,15 @@
+function RowOptions() {
+    return (
+        <button
+            className="purchases__content purchases__options"
+            aria-label="options"
+            aria-haspopup="true"
+        >
+            <span />
+            <span />
+            <span />
+        </button>
+    );
+};
+
+export default RowOptions;
